Fix iOS build: add new files to Xcode project and extract PasswordStrength
- Register BiometricService.swift, CertificatePinningService.swift,
  SkeletonView.swift, and Colors.swift in project.pbxproj so Xcode
  compiles them (fixes 'cannot find in scope' errors)
- Extract PasswordStrength enum from AuthView.swift into its own file
  to fix Swift type-checker timeout on the overly complex view body
- Add Components group under Views in Xcode project structure
